+++ src/models/PullRequestBase.ts
@@ -1,9 +1,0 @@
-export interface PullRequestBase {
-  title: string;
-  number: number;
-  repository: string;
-  url: string;
-  labels: string[];
-  author: string;
-  isDraft: boolean;
-}