+++ src/models/PullRequestInvolved.ts
@@ -1,3 +1,0 @@
-import { PullRequestBase } from "./PullRequestBase";
-
-export interface PullRequestInvolved extends PullRequestBase {}